+++ acme/resources/BAT.py
@@ -37,7 +37,7 @@ class BAT(MgmtObj):
 			'lt': None,
 			'et': None,
 			'lbl': None,
-			'hld': None,
+			'cstn': None,
 			'acpi':None,
 			'at': None,
 			'aa': None,