+++ acme/resources/CIN.py
@@ -29,7 +29,7 @@ class CIN(AnnounceableResource):
 			'lt': None,
 			'et': None,
 			'lbl': None,
-			'hld': None,
+			'cstn': None,
 			'at': None,
 			'aa': None,
 			'ast': None,