+++ acme/resources/CNT.py
@@ -34,7 +34,7 @@ class CNT(AnnounceableResource):
 			'lt': None,
 			'et': None,
 			'lbl': None,
-			'hld': None,
+			'cstn': None,
 			'acpi':None,
 			'at': None,
 			'aa': None,