+++ acme/resources/CSR.py
@@ -31,7 +31,7 @@ class CSR(AnnounceableResource):
 			'lt': None,
 			'et': None,
 			'lbl': None,
-			'hld': None,
+			'cstn': None,
 			'acpi':None,
 			'daci': None,
 			'at': None,