+++ acme/resources/DVC.py
@@ -26,7 +26,7 @@ class DVC(MgmtObj):
 			'lt': None,
 			'et': None,
 			'lbl': None,
-			'hld': None,
+			'cstn': None,
 			'acpi':None,
 			'at': None,
 			'aa': None,