+++ acme/resources/DVI.py
@@ -30,7 +30,7 @@ class DVI(MgmtObj):
 			'lt': None,
 			'et': None,
 			'lbl': None,
-			'hld': None,
+			'cstn': None,
 			'acpi':None,
 			'at': None,
 			'aa': None,