+++ acme/resources/EVL.py
@@ -40,7 +40,7 @@ class EVL(MgmtObj):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'at': None,
 		'aa': None,