+++ acme/resources/FCNT.py
@@ -35,7 +35,7 @@ class FCNT(AnnounceableResource):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'at': None,
 		'aa': None,