+++ acme/resources/FWR.py
@@ -36,7 +36,7 @@ class FWR(MgmtObj):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'at': None,
 		'aa': None,