+++ acme/resources/MEM.py
@@ -29,7 +29,7 @@ class MEM(MgmtObj):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'at': None,
 		'aa': None,