+++ acme/resources/NOD.py
@@ -37,7 +37,7 @@ class NOD(AnnounceableResource):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'at': None,
 		'aa': None,