+++ acme/resources/NYCFC.py
@@ -25,7 +25,7 @@ class NYCFC(MgmtObj):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'at': None,
 		'aa': None,