+++ acme/resources/SUB.py
@@ -33,7 +33,7 @@ class SUB(Resource):
 		'lt': None,
 		'et': None,
 		'lbl': None,
-		'hld': None,
+		'cstn': None,
 		'acpi':None,
 		'daci': None,
 		'cr': None,