+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSE] Adapted TimeSeries to latest R4 changes. 
 - [CSE] Changed the default release version to 4. Also, the supported and the actual release versions are now fully configurable (in the config file).
 - [CSE] Indicated Release Version must now be the highest value in the list of supported release versions.
+- [CSE] Changed name of *holder* attribute to *custodian* according to spec change.
 - [TESTS] Replaced CSE test cases' re-configurations with upper tester commands / script calls.
 
 ### Removed